Fix conversion compiler warnings
Using a release build with GCC 14.2.0 and -Wextra -Wconversion -Wall.
The generated binaries are not changed by this commit.

+++ taglib/dsdiff/dsdiffproperties.cpp
@@ -58,7 +58,7 @@ DSDIFF::Properties::Properties(unsigned int sampleRate,
   d->sampleRate = sampleRate;
   d->bitrate = bitrate;
   d->length = d->sampleRate > 0
-    ? static_cast<int>(d->sampleCount * 1000.0 / d->sampleRate + 0.5)
+    ? static_cast<int>(static_cast<double>(d->sampleCount) * 1000.0 / d->sampleRate + 0.5)
     : 0;
 }
